refactor(useDisplayedFolder: Extract from selectors to hooks
refacto(useDisplayedFolder): Use default
