+++ src/drive/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useCurrentFileId } from './useCurrentFileId'
 export { default as useCurrentFolderId } from './useCurrentFolderId'
+export { default as useDisplayedFolder } from './useDisplayedFolder'
