+++ src/drive/web/modules/drive/Toolbar/components/DownloadButtonItem.jsx
@@ -9,7 +9,7 @@ import Icon from 'cozy-ui/transpiled/react/Icon'
 import DownloadIcon from 'cozy-ui/transpiled/react/Icons/Download'
 
 import { downloadFiles } from 'drive/web/modules/actions/utils'
-import { useDisplayedFolder } from 'drive/web/modules/selectors'
+import { useDisplayedFolder } from 'drive/hooks'
 
 const DownloadButtonItem = ({ t, downloadAll }) => {
   const displayedFolder = useDisplayedFolder()
