+++ src/drive/web/modules/drive/Toolbar/components/InsideRegularFolder.jsx
@@ -1,7 +1,6 @@
 import { ROOT_DIR_ID } from 'drive/constants/config'
 import { isEncryptedFolder } from 'drive/lib/encryption'
-import { useDisplayedFolder } from 'drive/web/modules/selectors'
-import { useCurrentFolderId } from 'drive/hooks'
+import { useCurrentFolderId, useDisplayedFolder } from 'drive/hooks'
 
 /**
  * Displays its children only if we are in a normal folder (eg. not the root folder or a special view like sharings or recent)
