+++ src/drive/web/modules/drive/Toolbar/share/ShareItem.jsx
@@ -6,7 +6,7 @@ import { translate } from 'cozy-ui/transpiled/react/I18n'
 import { ActionMenuItem } from 'cozy-ui/transpiled/react/ActionMenu'
 import Icon from 'cozy-ui/transpiled/react/Icon'
 import ShareIcon from 'cozy-ui/transpiled/react/Icons/Share'
-import { useDisplayedFolder } from 'drive/web/modules/selectors'
+import { useDisplayedFolder } from 'drive/hooks'
 
 import styles from 'drive/styles/toolbar.styl'
 import shareContainer from './share'
