+++ src/drive/web/modules/drive/Toolbar/share/SharedRecipients.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { SharedRecipients } from 'cozy-sharing'
 import shareContainer from './share'
 
-import { useDisplayedFolder } from 'drive/web/modules/selectors'
+import { useDisplayedFolder } from 'drive/hooks'
 
 const SharedRecipientsComponent = ({ share }) => {
   const displayedFolder = useDisplayedFolder()
